kv: split out liveness persistence
The liveness code deals with both in-memory and persistent versions
internally. This change simplifies and makes the usage of the underlying
storage more consistent.

Epic: none

Release note: None
